+++ rpc/call_test.go
@@ -2,12 +2,12 @@ package rpc
 
 import (
 	"context"
-	"fmt"
 	"testing"
 
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/internal/tests"
 	internalUtils "github.com/NethermindEth/starknet.go/internal/utils"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -27,9 +27,9 @@ import (
 //
 //	none
 func TestCall(t *testing.T) {
-	tests.RunTestOn(t, tests.MockEnv, tests.TestnetEnv, tests.MainnetEnv, tests.DevnetEnv)
+	tests.RunTestOn(t, tests.MockEnv, tests.TestnetEnv, tests.MainnetEnv, tests.DevnetEnv, tests.IntegrationEnv)
 
-	testConfig := beforeEach(t, false)
+	testConfig := BeforeEach(t, false)
 
 	type testSetType struct {
 		name                  string
@@ -48,7 +48,7 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("decimals"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:               WithBlockTag("latest"),
+				BlockID:               WithBlockTag(BlockTagLatest),
 				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x12"),
 			},
 		},
@@ -56,23 +56,43 @@ func TestCall(t *testing.T) {
 			{
 				name: "Ok",
 				FunctionCall: FunctionCall{
-					ContractAddress:    internalUtils.TestHexToFelt(t, "0xdeadbeef"),
+					ContractAddress:    internalUtils.RANDOM_FELT,
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("decimals"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:               WithBlockTag("latest"),
-				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0xdeadbeef"),
+				BlockID:               WithBlockTag(BlockTagLatest),
+				ExpectedPatternResult: internalUtils.RANDOM_FELT,
 			},
 		},
 		tests.TestnetEnv: {
 			{
-				name: "Ok",
+				name: "Ok - latest block tag",
 				FunctionCall: FunctionCall{
 					ContractAddress:    internalUtils.TestHexToFelt(t, "0x025633c6142D9CA4126e3fD1D522Faa6e9f745144aba728c0B3FEE38170DF9e7"),
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:               WithBlockTag("latest"),
+				BlockID:               WithBlockTag(BlockTagLatest),
+				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x506f736974696f6e"),
+			},
+			{
+				name: "Ok - pre_confirmed block tag",
+				FunctionCall: FunctionCall{
+					ContractAddress:    internalUtils.TestHexToFelt(t, "0x025633c6142D9CA4126e3fD1D522Faa6e9f745144aba728c0B3FEE38170DF9e7"),
+					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
+					Calldata:           []*felt.Felt{},
+				},
+				BlockID:               WithBlockTag(BlockTagPre_confirmed),
+				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x506f736974696f6e"),
+			},
+			{
+				name: "Ok - l1_accepted block tag",
+				FunctionCall: FunctionCall{
+					ContractAddress:    internalUtils.TestHexToFelt(t, "0x025633c6142D9CA4126e3fD1D522Faa6e9f745144aba728c0B3FEE38170DF9e7"),
+					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
+					Calldata:           []*felt.Felt{},
+				},
+				BlockID:               WithBlockTag(BlockTagL1Accepted),
 				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x506f736974696f6e"),
 			},
 			{
@@ -82,7 +102,7 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
 					Calldata:           []*felt.Felt{&felt.Zero},
 				},
-				BlockID:       WithBlockTag("latest"),
+				BlockID:       WithBlockTag(BlockTagLatest),
 				ExpectedError: ErrContractError,
 			},
 			{
@@ -92,7 +112,7 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("RANDOM_STRINGGG"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:       WithBlockTag("latest"),
+				BlockID:       WithBlockTag(BlockTagLatest),
 				ExpectedError: ErrEntrypointNotFound,
 			},
 			{
@@ -102,7 +122,7 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:       WithBlockNumber(9999999999999999999),
+				BlockID:       WithBlockNumber(999999999999999),
 				ExpectedError: ErrBlockNotFound,
 			},
 			{
@@ -112,8 +132,20 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("name"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:       WithBlockTag("latest"),
+				BlockID:       WithBlockTag(BlockTagLatest),
 				ExpectedError: ErrContractNotFound,
+			},
+		},
+		tests.IntegrationEnv: {
+			{
+				name: "Ok",
+				FunctionCall: FunctionCall{
+					ContractAddress:    internalUtils.TestHexToFelt(t, "0x04718f5a0fc34cc1af16a1cdee98ffb20c31f5cd61d6ab07201858f4287c938d"),
+					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("decimals"),
+					Calldata:           []*felt.Felt{},
+				},
+				BlockID:               WithBlockTag(BlockTagLatest),
+				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x12"),
 			},
 		},
 		tests.MainnetEnv: {
@@ -124,23 +156,23 @@ func TestCall(t *testing.T) {
 					EntryPointSelector: internalUtils.GetSelectorFromNameFelt("decimals"),
 					Calldata:           []*felt.Felt{},
 				},
-				BlockID:               WithBlockTag("latest"),
+				BlockID:               WithBlockTag(BlockTagLatest),
 				ExpectedPatternResult: internalUtils.TestHexToFelt(t, "0x12"),
 			},
 		},
 	}[tests.TEST_ENV]
 
 	for _, test := range testSet {
-		t.Run(fmt.Sprintf("Network: %s, Test: %s", tests.TEST_ENV, test.name), func(t *testing.T) {
-			output, err := testConfig.provider.Call(context.Background(), test.FunctionCall, test.BlockID)
-			if err != nil {
+		t.Run("Test: "+test.name, func(t *testing.T) {
+			output, err := testConfig.Provider.Call(context.Background(), test.FunctionCall, test.BlockID)
+			if test.ExpectedError != nil {
 				rpcErr, ok := err.(*RPCError)
 				require.True(t, ok)
-				require.ErrorContains(t, test.ExpectedError, rpcErr.Message)
+				assert.ErrorContains(t, test.ExpectedError, rpcErr.Message)
 			} else {
 				require.NoError(t, err)
 				require.NotEmpty(t, output, "should return an output")
-				require.Equal(t, test.ExpectedPatternResult, output[0])
+				assert.Equal(t, test.ExpectedPatternResult, output[0])
 			}
 		})
 	}