+++ rpc/trace.go
@@ -69,12 +69,17 @@ func (provider *Provider) TraceTransaction(ctx context.Context, transactionHash 
 //
 // Parameters:
 //   - ctx: the context.Context object for controlling the request
-//   - blockHash: the hash of the block to retrieve the traces from
+//   - blockID: the block to retrieve the traces from. `pre_confirmed` tag is not allowed
 //
 // Returns:
 //   - []Trace: a slice of Trace objects representing the traces of transactions in the block
 //   - error: an error if there was a problem retrieving the traces.
 func (provider *Provider) TraceBlockTransactions(ctx context.Context, blockID BlockID) ([]Trace, error) {
+	err := checkForPre_confirmed(blockID)
+	if err != nil {
+		return nil, err
+	}
+
 	var output []Trace
 	if err := do(ctx, provider.c, "starknet_traceBlockTransactions", &output, blockID); err != nil {
 		return nil, tryUnwrapToRPCErr(err, ErrBlockNotFound)