+++ rpc/types_block.go
@@ -20,8 +20,8 @@ type Block struct {
 	Transactions []BlockTransaction `json:"transactions"`
 }
 
-type PendingBlock struct {
-	PendingBlockHeader
+type Pre_confirmedBlock struct {
+	Pre_confirmedBlockHeader
 	Transactions []BlockTransaction `json:"transactions"`
 }
 
@@ -42,8 +42,8 @@ type TransactionWithReceipt struct {
 }
 
 // The dynamic block being constructed by the sequencer. Note that this object will be deprecated upon decentralisation.
-type PendingBlockWithReceipts struct {
-	PendingBlockHeader
+type Pre_confirmedBlockWithReceipts struct {
+	Pre_confirmedBlockHeader
 	BlockBodyWithReceipts
 }
 
@@ -54,8 +54,8 @@ type BlockTxHashes struct {
 	Transactions []*felt.Felt `json:"transactions"`
 }
 
-type PendingBlockTxHashes struct {
-	PendingBlockHeader
+type Pre_confirmedBlockTxHashes struct {
+	Pre_confirmedBlockHeader
 	Transactions []*felt.Felt `json:"transactions"`
 }
 
@@ -84,9 +84,10 @@ type BlockHeader struct {
 	StarknetVersion string `json:"starknet_version"`
 }
 
-type PendingBlockHeader struct {
-	// ParentHash The hash of this block's parent
-	ParentHash *felt.Felt `json:"parent_hash"`
+type Pre_confirmedBlockHeader struct {
+	// The block number of the block that the proposer is currently building.
+	// Note that this is a local view of the node, whose accuracy depends on its polling interval length.
+	Number uint64 `json:"block_number"`
 	// Timestamp the time in which the block was created, encoded in Unix time
 	Timestamp uint64 `json:"timestamp"`
 	// SequencerAddress the StarkNet identity of the sequencer submitting this block
@@ -113,10 +114,12 @@ type BlockHashAndNumberOutput struct {
 type BlockTag string
 
 const (
-	// BlockTagLatest represents the latest confirmed block.
+	// The block which is currently being built by the block proposer in height `latest` + 1.
+	BlockTagPre_confirmed BlockTag = "pre_confirmed"
+	// The latest Starknet block finalised by the consensus on L2.
 	BlockTagLatest BlockTag = "latest"
-	// BlockTagPending represents the pending block that is yet to be confirmed.
-	BlockTagPending BlockTag = "pending"
+	// The latest Starknet block which was included in a state update on L1 and finalised by the consensus on L1.
+	BlockTagL1Accepted BlockTag = "l1_accepted"
 )
 
 // BlockID is a struct that is used to choose between different
@@ -124,15 +127,81 @@ const (
 type BlockID struct {
 	Number *uint64    `json:"block_number,omitempty"`
 	Hash   *felt.Felt `json:"block_hash,omitempty"`
-	Tag    BlockTag   `json:"block_tag,omitempty"`
+	// A tag specifying a dynamic reference to a block. Tag `l1_accepted` refers
+	// to the latest Starknet block which was included in a state update on L1 and
+	// finalised by the consensus on L1. Tag `latest` refers to the latest Starknet
+	// block finalised by the consensus on L2. Tag `pre_confirmed` refers to the block
+	// which is currently being built by the block proposer in height `latest` + 1.
+	Tag BlockTag `json:",omitempty"`
 }
 
-// checkForPending checks if the block ID has the 'pending' tag. If it does, it returns an error.
-// This is used to prevent the user from using the 'pending' tag on methods that do not support it.
-func checkForPending(b BlockID) error {
-	if b.Tag == BlockTagPending {
-		return errors.Join(ErrInvalidBlockID, errors.New("'pending' tag is not supported on this method"))
+// Block hash, number or tag, same as BLOCK_ID, but without 'pre_confirmed' or 'l1_accepted'
+type SubscriptionBlockID BlockID
+
+// BlockID returns a BlockID from a SubscriptionBlockID.
+func (b *SubscriptionBlockID) BlockID() BlockID {
+	return BlockID{
+		Number: b.Number,
+		Hash:   b.Hash,
+		Tag:    b.Tag,
 	}
+}
+
+// WithBlockNumber sets the block number for the SubscriptionBlockID.
+func (b *SubscriptionBlockID) WithBlockNumber(number uint64) SubscriptionBlockID {
+	b.Number = &number
+
+	return *b
+}
+
+// WithBlockHash sets the block hash for the SubscriptionBlockID.
+func (b *SubscriptionBlockID) WithBlockHash(hash *felt.Felt) SubscriptionBlockID {
+	b.Hash = hash
+
+	return *b
+}
+
+// WithLatestTag sets the block tag to latest for the SubscriptionBlockID.
+// It's the only block tag allowed for this type.
+func (b *SubscriptionBlockID) WithLatestTag() SubscriptionBlockID {
+	b.Tag = BlockTagLatest
+
+	return *b
+}
+
+func (b *BlockID) UnmarshalJSON(data []byte) error {
+	var tag string
+
+	if err := json.Unmarshal(data, &tag); err == nil {
+		if tag == string(BlockTagPre_confirmed) || tag == string(BlockTagLatest) || tag == string(BlockTagL1Accepted) {
+			b.Tag = BlockTag(tag)
+
+			return nil
+		}
+	}
+
+	type Alias BlockID
+	var aux Alias
+	if err := json.Unmarshal(data, &aux); err == nil {
+		*b = BlockID(aux)
+
+		return nil
+	}
+
+	return errors.New("invalid block ID")
+}
+
+func (b *SubscriptionBlockID) UnmarshalJSON(data []byte) error {
+	var aux BlockID
+	if err := json.Unmarshal(data, &aux); err != nil {
+		return err
+	}
+
+	if aux.Tag == BlockTagPre_confirmed || aux.Tag == BlockTagL1Accepted {
+		return fmt.Errorf("invalid block tag for this type: %s", aux.Tag)
+	}
+
+	*b = SubscriptionBlockID(aux)
 
 	return nil
 }
@@ -150,7 +219,7 @@ func checkForPending(b BlockID) error {
 //   - []byte: the JSON representation of the BlockID
 //   - error: any error that occurred during the marshalling process
 func (b BlockID) MarshalJSON() ([]byte, error) {
-	if b.Tag == BlockTagPending || b.Tag == BlockTagLatest {
+	if b.Tag == BlockTagPre_confirmed || b.Tag == BlockTagLatest || b.Tag == BlockTagL1Accepted {
 		return []byte(strconv.Quote(string(b.Tag))), nil
 	}
 
@@ -169,13 +238,30 @@ func (b BlockID) MarshalJSON() ([]byte, error) {
 	return json.Marshal(nil)
 }
 
+func (b SubscriptionBlockID) MarshalJSON() ([]byte, error) {
+	if b.Tag == BlockTagPre_confirmed || b.Tag == BlockTagL1Accepted {
+		return nil, fmt.Errorf("invalid block tag for this type: %s", b.Tag)
+	}
+
+	return BlockID(b).MarshalJSON()
+}
+
+// checkForPre_confirmed checks if the block ID has the 'pre_confirmed' tag. If it does, it returns an error.
+// This is used to prevent the user from using the 'pre_confirmed' tag on methods that do not support it.
+func checkForPre_confirmed(b BlockID) error {
+	if b.Tag == BlockTagPre_confirmed {
+		return errors.Join(ErrInvalidBlockID, errors.New("'pre_confirmed' tag is not supported on this method"))
+	}
+
+	return nil
+}
+
 type BlockStatus string
 
 const (
-	BlockStatus_Pending      BlockStatus = "PENDING"
-	BlockStatus_AcceptedOnL2 BlockStatus = "ACCEPTED_ON_L2"
-	BlockStatus_AcceptedOnL1 BlockStatus = "ACCEPTED_ON_L1"
-	BlockStatus_Rejected     BlockStatus = "REJECTED"
+	BlockStatus_Pre_confirmed BlockStatus = "PRE_CONFIRMED"
+	BlockStatus_AcceptedOnL2  BlockStatus = "ACCEPTED_ON_L2"
+	BlockStatus_AcceptedOnL1  BlockStatus = "ACCEPTED_ON_L1"
 )
 
 // UnmarshalJSON unmarshals the JSON representation of a BlockStatus.
@@ -195,14 +281,12 @@ func (bs *BlockStatus) UnmarshalJSON(data []byte) error {
 	}
 
 	switch unquoted {
-	case "PENDING":
-		*bs = BlockStatus_Pending
+	case "PRE_CONFIRMED":
+		*bs = BlockStatus_Pre_confirmed
 	case "ACCEPTED_ON_L2":
 		*bs = BlockStatus_AcceptedOnL2
 	case "ACCEPTED_ON_L1":
 		*bs = BlockStatus_AcceptedOnL1
-	case "REJECTED":
-		*bs = BlockStatus_Rejected
 	default:
 		return fmt.Errorf("unsupported status: %s", data)
 	}